+++ src/models/report_views/productivity_balance.py
@@ -7,5 +7,5 @@ __all__ = ('UnitProductivityBalanceStatisticsViewDTO',)
 class UnitProductivityBalanceStatisticsViewDTO:
     unit_name: str
     sales_per_labor_hour: int
-    orders_per_labor_hour: int
+    orders_per_labor_hour: float
     stop_sale_duration_in_seconds: int
